+++ examples/passport/telegram-mini-app/app/components/Home.tsx
@@ -11,8 +11,8 @@ import {
 	Heading,
 	Stack,
 } from "@biom3/react";
-import { Connect } from "./components/Connect";
-import { TransferAsset } from "./components/TransferAsset";
+import { Connect } from "./Connect";
+import { TransferAsset } from "./TransferAsset";
 
 export default function Home() {
 	const [walletAddress, setWalletAddress] = useState<string>("");
